testing/web-platform/tests/web-animations/keyframe-effect/getComputedTiming.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>KeyframeEffectReadOnly getComputedTiming() tests</title>
<link rel="help" href="https://w3c.github.io/web-animations/#dom-animationeffectreadonly-getcomputedtiming">
<link rel="author" title="Boris Chiou" href="mailto:boris.chiou@gmail.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<div id="target"></div>
<script>
"use strict";

var target = document.getElementById("target");

test(function(t) {
  var effect = new KeyframeEffectReadOnly(target,
                                          { left: ["10px", "20px"] });

  var ct = effect.getComputedTiming();
  assert_equals(ct.delay, 0, "computed delay");
  assert_equals(ct.fill, "none", "computed fill");
  assert_equals(ct.iterations, 1.0, "computed iterations");
  assert_equals(ct.duration, 0, "computed duration");
  assert_equals(ct.direction, "normal", "computed direction");
}, "values of getComputedTiming() when a KeyframeEffectReadOnly is " +
   "constructed without any KeyframeEffectOptions object");

var gGetComputedTimingTests = [
  { desc:     "an empty KeyframeEffectOptions object",
    input:    { },
    expected: { } },
  { desc:     "a normal KeyframeEffectOptions object",
    input:    { delay: 1000,
                fill: "auto",
                iterations: 5.5,
                duration: "auto",
                direction: "alternate" },
    expected: { delay: 1000,
                fill: "none",
                iterations: 5.5,
                duration: 0,
                direction: "alternate" } },
  { desc:     "a double value",
    input:    3000,
    timing:   { duration: 3000 },
    expected: { delay: 0,
                fill: "none",
                iterations: 1,
                duration: 3000,
                direction: "normal" } },
  { desc:     "+Infinity",
    input:    Infinity,
    expected: { duration: Infinity } },
  { desc:     "-Infinity",
    input:    -Infinity,
    expected: { duration: 0 } },
  { desc:     "NaN",
    input:    NaN,
    expected: { duration: 0 } },
  { desc:     "a negative value",
    input:    -1,
    expected: { duration: 0 } },
  { desc:     "an Infinity duration",
    input:    { duration: Infinity },
    expected: { duration: Infinity } },
  { desc:     "a negative Infinity duration",
    input:    { duration: -Infinity },
    expected: { duration: 0 } },
  { desc:     "a NaN duration",
    input:    { duration: NaN },
    expected: { duration: 0 } },
  { desc:     "a negative duration",
    input:    { duration: -1},
    expected: { duration: 0 } },
  { desc:     "a string duration",
    input:    { duration: "merrychristmas"},
    expected: { duration: 0 } },
  { desc:     "an auto duration",
    input:    { duration: "auto" },
    expected: { duration: 0 } },
  { desc:     "an Infinity iterations",
    input:    { iterations: Infinity },
    expected: { iterations: Infinity } },
  { desc:     "a negative Infinity iterations",
    input:    { iterations: -Infinity},
    expected: { iterations: 1 } },
  { desc:     "a NaN iterations",
    input:    { iterations: NaN },
    expected: { iterations: 1 } },
  { desc:     "a negative iterations",
    input:    { iterations: -1 },
    expected: { iterations: 1 } },
  { desc:     "an auto fill",
    input:    { fill: "auto" },
    expected: { fill: "none" } },
  { desc:     "a forwards fill",
    input:    { fill: "forwards" },
    expected: { fill: "forwards" } }
];

gGetComputedTimingTests.forEach(function(stest) {
  test(function(t) {
    var effect = new KeyframeEffectReadOnly(target,
                                            { left: ["10px", "20px"] },
                                            stest.input);

    // Helper function to provide default expected values when the test does
    // not supply them.
    var expected = function(field, defaultValue) {
      return field in stest.expected ? stest.expected[field] : defaultValue;
    };

    var ct = effect.getComputedTiming();
    assert_equals(ct.delay, expected("delay", 0),
                  "computed delay");
    assert_equals(ct.fill, expected("fill", "none"),
                  "computed fill");
    assert_equals(ct.iterations, expected("iterations", 1),
                  "computed iterations");
    assert_equals(ct.duration, expected("duration", 0),
                  "computed duration");
    assert_equals(ct.direction, expected("direction", "normal"),
                  "computed direction");

  }, "values of getComputedTiming() when a KeyframeEffectReadOnly is " +
     "constructed by " + stest.desc);
});

done();
</script>
</body>
